# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AgentPilot API
  description: >
    Copiloto de conocimiento RAG para agentes de Contact Center.
    Contrato diseñado contract-first (Spec Driven Development): este fichero
    es la fuente de verdad de la API; el código se implementa contra él.


    **Campañas (v1.1.0):** toda la documentación pertenece a una campaña y el
    asistente responde **únicamente** con el corpus de la campaña seleccionada.
    Por eso `campaignId` es obligatorio en `POST /chat/ask` y en la subida de
    documentos: es un cambio incompatible respecto a la 1.0.0, deliberado, porque
    un valor por defecto silencioso convertiría un olvido del cliente en una fuga
    de información entre campañas.


    **Prompt por capas (v1.2.0):** las instrucciones propias de una campaña se
    gestionan aparte, en `/campaigns/{campaignId}/prompt`, y ya no viajan dentro
    de `Campaign`/`CampaignRequest` (cambio incompatible respecto a la 1.1.0).
    El asistente compone siempre tres capas — núcleo inmutable (código) + bloque
    de campaña (dato) + reafirmación del núcleo (código) — y esa composición es
    la misma tanto para el chat real como para `POST .../prompt/preview`, para
    que lo que se previsualiza sea exactamente lo que se publica. Cada cambio
    (incluida una restauración) añade una entrada a `/prompt/versions`; ninguna
    se sobrescribe ni se borra.
  version: 1.2.0
  contact:
    name: Camilo Rodríguez
    url: https://github.com/camilorkll/AgentPilot
servers:
...
paths:
...
  /campaigns:
    get:
      tags: [campaigns]
      operationId: listCampaigns
      summary: Lista todas las campañas con su estado y volumen de documentación
      description: Requiere rol `admin`.
      responses:
        "200":
          description: Listado de campañas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [campaigns]
      operationId: createCampaign
      summary: Crea una campaña
      description: >
        Requiere rol `admin`. La campaña nace **activa**. El nombre es único
        (sin distinguir mayúsculas).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignRequest"
      responses:
        "201":
          description: Campaña creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Ya existe una campaña con ese nombre
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /campaigns/active:
    get:
      tags: [campaigns]
      operationId: listActiveCampaigns
      summary: Campañas activas sobre las que un agente puede trabajar
      description: >
        Alimenta el selector de campaña del agente. Devuelve **solo** las
        campañas en estado `active`, con una proyección reducida: el agente no
        necesita (ni debe ver) las instrucciones del asistente.
      responses:
        "200":
          description: Campañas disponibles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /campaigns/{campaignId}:
    get:
      tags: [campaigns]
      operationId: getCampaign
      summary: Consulta una campaña
      description: Requiere rol `admin`.
      parameters:
        - $ref: "#/components/parameters/CampaignIdPath"
      responses:
        "200":
          description: Campaña
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [campaigns]
      operationId: updateCampaign
      summary: Modifica el nombre y las instrucciones del asistente
      description: >
        Requiere rol `admin`. **No cambia el estado**: para eso está
        `POST /campaigns/{campaignId}/status`, porque cerrar una campaña la
        congela y no debe poder ocurrir como efecto secundario de renombrarla.
        Una campaña `closed` es de solo lectura y responde 409.
      parameters:
        - $ref: "#/components/parameters/CampaignIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignRequest"
      responses:
        "200":
          description: Campaña actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: La campaña está cerrada, o el nombre ya existe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [campaigns]
      operationId: deleteCampaign
      summary: Elimina una campaña y todo su corpus
      description: >
        Requiere rol `admin`. **Solo se permite si la campaña está `closed`**;
        en cualquier otro estado responde 409. Borra en cascada sus documentos y
        fragmentos indexados. Las conversaciones y la telemetría se conservan
        (dejan de apuntar a la campaña pero mantienen su nombre para los
        informes). Es irreversible.
      parameters:
        - $ref: "#/components/parameters/CampaignIdPath"
      responses:
        "204":
          description: Campaña y corpus eliminados
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: La campaña no está cerrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /campaigns/{campaignId}/status:
    post:
      tags: [campaigns]
      operationId: setCampaignStatus
      summary: Cambia el estado de una campaña
      description: >
        Requiere rol `admin`. Las transiciones válidas son:


        - `active` ⇄ `inactive`

        - `inactive` → `closed` (solo se cierra una campaña ya inactiva)

        - `closed` → `inactive` (reapertura)


        Cualquier otra transición responde 409. Una campaña `closed` conserva
        todo y sus informes siguen consultables, pero no admite cambios en su
        documentación.
      parameters:
        - $ref: "#/components/parameters/CampaignIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignStatusRequest"
      responses:
        "200":
          description: Estado actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Transición no permitida desde el estado actual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /campaigns/{campaignId}/prompt:
    get:
      tags: [campaigns]
      operationId: getCampaignPrompt
      summary: Instrucciones vigentes de la campaña para el asistente
      description: >
        Requiere rol `admin`. Un objeto con todos los campos vacíos
        (`isEmpty: true`) significa que la campaña no tiene instrucciones
        propias: el asistente responde solo con el núcleo.
      parameters:
        - $ref: "#/components/parameters/CampaignIdPath"
      responses:
        "200":
          description: Instrucciones vigentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantPromptSettings"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [campaigns]
      operationId: updateCampaignPrompt
      summary: Publica unas instrucciones nuevas y añade una entrada al historial
      description: >
        Requiere rol `admin`. Enviar el formulario completamente vacío equivale
        a restaurar el comportamiento por defecto (solo el núcleo). Cada
        llamada crea una entrada nueva en `/prompt/versions`; ninguna entrada
        anterior se sobrescribe ni se borra. Una campaña `closed` es de solo
        lectura y responde 409.
      parameters:
        - $ref: "#/components/parameters/CampaignIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantPromptRequest"
      responses:
        "200":
          description: Instrucciones publicadas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptUpdateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/CampaignClosed"
  /campaigns/{campaignId}/prompt/versions:
    get:
      tags: [campaigns]
      operationId: listCampaignPromptVersions
      summary: Historial de instrucciones de la campaña, más reciente primero
      description: >
        Requiere rol `admin`. Registro de auditoría de solo lectura: incluye
        también las restauraciones, cada una como su propia entrada.
      parameters:
        - $ref: "#/components/parameters/CampaignIdPath"
      responses:
        "200":
          description: Historial de la campaña
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromptVersion"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /campaigns/{campaignId}/prompt/versions/{versionId}/restore:
    post:
      tags: [campaigns]
      operationId: restoreCampaignPromptVersion
      summary: Vuelve a aplicar una versión pasada del historial
      description: >
        Requiere rol `admin`. Restaurar es publicar de nuevo: crea una entrada
        de historial propia, no reescribe ni borra la versión de la que parte.
        Una campaña `closed` es de solo lectura y responde 409.
      parameters:
        - $ref: "#/components/parameters/CampaignIdPath"
        - $ref: "#/components/parameters/VersionIdPath"
      responses:
        "200":
          description: Instrucciones restauradas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptUpdateResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: >
            La campaña no existe, o la versión no existe o no pertenece a
            esta campaña
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          $ref: "#/components/responses/CampaignClosed"
  /campaigns/{campaignId}/prompt/preview:
    post:
      tags: [campaigns]
      operationId: previewCampaignPrompt
      summary: Compara la respuesta con las instrucciones publicadas y con un candidato sin guardar
      description: >
        Requiere rol `admin`. Recupera el contexto **una sola vez** para la
        pregunta de prueba y genera dos respuestas sobre él, para que la
        comparación sea justa. No exige que la campaña esté activa (es de
        solo lectura) y, a propósito, **no crea conversación ni telemetría**:
        no es tráfico real, es una herramienta de administración y no debe
        aparecer en el dashboard de coste ni en el historial de un agente.
      parameters:
        - $ref: "#/components/parameters/CampaignIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromptPreviewRequest"
      responses:
        "200":
          description: Respuestas comparadas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptPreviewResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CampaignIdPath:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VersionIdPath:
      name: versionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Petición inválida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Sin autenticar o token inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Autenticado pero sin permisos suficientes
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    CampaignClosed:
      description: >
        La campaña está cerrada y su documentación es de solo lectura
        (`code: campaign_closed`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    CampaignStatus:
      type: string
      description: >
        Ciclo de vida de una campaña. Se persiste como entero
        (`inactive` = 0, `active` = 1, `closed` = 2) y se expone con nombre para
        que el contrato se lea sin tabla de conversión.


        - `active`: visible para los agentes y respondiendo consultas.

        - `inactive`: retirada del selector; su documentación aún se puede editar.

        - `closed`: solo lectura. Conserva todo y sus informes siguen
        consultables, pero no admite cambios; es el único estado desde el que se
        puede eliminar.
      enum: [inactive, active, closed]
    Campaign:
      type: object
      required: [id, name, status, documentCount, createdAtUtc]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, maxLength: 120 }
        status:
          $ref: "#/components/schemas/CampaignStatus"
        documentCount:
          type: integer
          description: Documentos de la campaña, en cualquier estado de ingesta
        activeDocumentCount:
          type: integer
          description: Documentos indexados y activos, es decir consultables por el asistente
        closedAtUtc: { type: string, format: date-time, nullable: true }
        createdAtUtc: { type: string, format: date-time }
    CampaignSummary:
      type: object
      description: Proyección para el selector del agente, sin datos de configuración.
      required: [id, name, activeDocumentCount]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        activeDocumentCount:
          type: integer
          description: >
            Documentos consultables. A cero, el asistente solo podrá abstenerse,
            y conviene que el agente lo sepa antes de preguntar.
    CampaignRequest:
      type: object
      description: >
        Solo el nombre: las instrucciones del asistente se gestionan aparte,
        en `/campaigns/{campaignId}/prompt`.
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
    CampaignStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/CampaignStatus"
    AssistantPromptRequest:
      type: object
      description: >
        Instrucciones propias de una campaña para el asistente: negocio (tono,
        avisos, vocabulario), nunca reglas del sistema. Todos los campos son
        opcionales; un objeto completamente vacío equivale a restaurar el
        comportamiento por defecto (solo el núcleo).
      properties:
        tone:
          type: string
          enum: [cercano, neutro, formal]
          nullable: true
        detailLevel:
          type: string
          enum: [breve, normal, detallado]
          nullable: true
        mandatoryNotice:
          type: string
          maxLength: 300
          nullable: true
          description: Aviso que el asistente debe recordar siempre (p. ej. verificar identidad)
        avoidWords:
          type: array
          maxItems: 20
          items: { type: string, maxLength: 60 }
        extraInstructions:
          type: string
          maxLength: 2000
          nullable: true
          description: Texto libre acotado; único campo sin estructura del formulario
    AssistantPromptSettings:
      description: >
        Instrucciones vigentes de la campaña, tal como las devuelve la API
        (a diferencia de AssistantPromptRequest, aquí avoidWords e isEmpty
        siempre están presentes).
      type: object
      required: [avoidWords, isEmpty]
      properties:
        tone:
          type: string
          enum: [cercano, neutro, formal]
          nullable: true
        detailLevel:
          type: string
          enum: [breve, normal, detallado]
          nullable: true
        mandatoryNotice:
          type: string
          maxLength: 300
          nullable: true
        avoidWords:
          type: array
          maxItems: 20
          items: { type: string, maxLength: 60 }
        extraInstructions:
          type: string
          maxLength: 2000
          nullable: true
        isEmpty:
          type: boolean
          description: >
            True si ningún campo está informado: el asistente responde
            solo con el núcleo, igual que antes de que existiera esta función.
    PromptUpdateResponse:
      type: object
      required: [prompt, warnings, versionId, createdAtUtc]
      properties:
        prompt:
          $ref: "#/components/schemas/AssistantPromptSettings"
        warnings:
          type: array
          items: { type: string }
          description: >
            Avisos de lint no bloqueantes: patrones en el texto libre o el
            aviso que sugieren un intento de anular el núcleo (p. ej.
            "ignora", "responde siempre"). Se publica igualmente: el núcleo
            se reafirma después de este bloque de todos modos.
        versionId: { type: string, format: uuid }
        createdAtUtc: { type: string, format: date-time }
    PromptVersion:
      type: object
      required: [id, prompt, publishedBy, createdAtUtc]
      properties:
        id: { type: string, format: uuid }
        prompt:
          $ref: "#/components/schemas/AssistantPromptSettings"
        publishedBy:
          type: string
          description: Usuario autenticado que publicó este cambio (o restauración)
        createdAtUtc: { type: string, format: date-time }
    PromptPreviewRequest:
      allOf:
        - $ref: "#/components/schemas/AssistantPromptRequest"
        - type: object
          required: [question]
          properties:
            question:
              type: string
              minLength: 1
              description: Pregunta de prueba, resuelta contra el corpus real de la campaña
    PromptPreviewResponse:
      type: object
      required: [currentAnswer, candidateAnswer, citations, warnings]
      properties:
        currentAnswer:
          type: string
          description: Respuesta con las instrucciones publicadas actualmente
        candidateAnswer:
          type: string
          description: Respuesta con el candidato del cuerpo de la petición, sin guardar
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
          description: >
            Mismo contexto recuperado para ambas respuestas: la comparación es
            sobre el prompt, no sobre qué encontró la búsqueda
        warnings:
          type: array
          items: { type: string }
...
    Citation:
      type: object
      required: [documentId, documentTitle, chunkId, snippet, score]
      properties:
        documentId: { type: string, format: uuid }
        documentTitle: { type: string }
        chunkId: { type: string, format: uuid }
        snippet:
          type: string
          description: Fragmento del documento fuente usado en la respuesta
        score:
          type: number
          format: double
          description: Similitud del retrieval (0..1)
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: validation_error }
        message: { type: string }
        details:
          type: array
          items: { type: string }
          nullable: true
security:
  - bearerAuth: []